feat: Add optional `autocluster` user setting + Fix `yarn process:autocluster` command + Update example configuration file

+++ processing/processing/actions/d_autocluster.py
@@ -5,47 +5,52 @@ from processing.clusterings.AutoConsensusClustering import \
 from processing.common.Env import Env
 from processing.storage.Storage import Storage
 from processing.utils.print_new_line import print_new_line
+from processing.utils.print_skip_line import print_skip_line
 
 env = Env()
 storage = Storage(path=env.storage)
+autocluster = storage.get_autocluster()
 
-bands = storage.get_bands()
-integrations = storage.get_integrations_seconds()
+if not autocluster:
+    print_skip_line()
+else:
+    bands = storage.get_bands()
+    integrations = storage.get_integrations_seconds()
 
-print_new_line()
-print('AutoCluster loading')
+    for band in bands:
+        print_new_line()
+        print('AutoCluster loading')
 
-storage.delete_autocluster()
+        storage.delete_autocluster()
 
-for band in bands:
-    for integration in integrations:
-        grouped_features = storage.read_grouped_features_all_files(
-            band=band,
-            integration=integration,
-            unwrap=True,
-        )
+        for integration in integrations:
+            grouped_features = storage.read_grouped_features_all_files(
+                band=band,
+                integration=integration,
+                unwrap=True,
+            )
 
-        clustering = AutoConsensusClustering(
-            features=grouped_features,
-            iterations=100,
-            min_cluster_size=20,
-            max_cluster_size=60,
-            threshold=0.9,
-        )
+            clustering = AutoConsensusClustering(
+                features=grouped_features,
+                iterations=100,
+                min_cluster_size=20,
+                max_cluster_size=60,
+                threshold=0.9,
+            )
 
-        consensus = clustering.get_consensus()
-        score = clustering.get_score()
-        numpy.set_printoptions(threshold=numpy.inf)
+            consensus = clustering.get_consensus()
+            score = clustering.get_score()
+            numpy.set_printoptions(threshold=numpy.inf)
 
-        print('Consensus:')
+            print('Consensus:')
 
-        for value in list(set(consensus)):
-            print(f'  {value}: {consensus.count(value)}')
+            for value in list(set(consensus)):
+                print(f'  {value}: {consensus.count(value)}')
 
-        print(f'Score: {score}')
+            print(f'Score: {score}')
 
-        storage.write_autocluster(
-            autocluster=consensus,
-            band=band,
-            integration=integration,
-        )
+            storage.write_autocluster(
+                autocluster=consensus,
+                band=band,
+                integration=integration,
+            )
